service-workers/service-worker/fetch-request-css-base-url.https.html

<!DOCTYPE html>
<title>Service Worker: CSS's base URL must be the request URL even when fetched from other URL</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/test-helpers.sub.js?pipe=sub"></script>
<script src="resources/service-worker-recorder.js"></script>
<script>
// Continually execute the provided function until it returns the value `true`
// or a Promise that resolves to `true`.
function waitFor(condition) {
  function go() {
      return Promise.resolve(condition())
        .then(function(result) {
            if (result === true) {
              return;
            }
            return new Promise(function(resolve) { step_timeout(resolve, 100); })
                .then(go);
          });
    };

  return go();
};

async_test(function(t) {
    var SCOPE = 'resources/fetch-request-css-base-url-iframe.html';
    var SCRIPT = 'resources/fetch-request-css-base-url-worker.js';
    var worker;
    var testDonePromise;

    return service_worker_unregister_and_register(t, SCRIPT, SCOPE)
      .then(function(registration) {
          worker = registration.installing;
          return wait_for_state(t, worker, 'activated')
            .then(function() {
                return ServiceWorkerRecorder.client.clear(registration.active);
              });
        })
      .then(function() { return with_iframe(SCOPE); })
      .then(function(iframe) {
          var requestInfo;
          // There is no event that signals when a stylesheet's sub-resources
          // have been loaded, so a polling approach must be used to delay test
          // execution until after the service worker has intercepted the
          // request as expected.
          return waitFor(function() {
              return ServiceWorkerRecorder.client.read(worker)
                .then(function(results) {
                    requestInfo = results[0];
                    return results.length > 0;
                  });
            })
            .then(function() {
                iframe.remove();
                return requestInfo;
              });
        })
      .then(function(result) {
          var base = get_host_info()['HTTPS_ORIGIN'] + base_path();
          assert_equals(
            result.url,
            base + 'resources/dummy.png',
            'The base URL while loading the images referred from CSS ' +
            'must be the request URL of CSS.');
          assert_equals(
            result.referrer,
            base + 'resources/fetch-request-css-base-url-style.css',
            'While loading the image defined in CSS the referrer must ' +
            'be the request URL of CSS.');
          return service_worker_unregister_and_done(t, SCOPE);
        })
      .catch(unreached_rejection(t));
  }, 'CSS\'s base URL must be the request URL even when fetched from other URL.');
</script>
